+++ parser/parser_aead.c
@@ -173,7 +173,84 @@ static int aead_tester(struct json_object *in, struct json_object *out,
 	const struct json_array aead_testanchor = SET_ARRAY(aead_testanchor_entries, NULL);
 
 	/* Process all. */
-	return process_json(&aead_testanchor, "0.5", in, out);
+	return process_json(&aead_testanchor, "1.0", in, out);
+}
+
+static int aead_siv_tester(struct json_object *in, struct json_object *out,
+			   struct aead_data *vector,
+			   const struct json_callbacks *callbacks)
+{
+	if (!aead_backend) {
+		logger(LOGGER_ERR,
+		       "No backend implementation for AEAD ciphers available\n");
+		return EOPNOTSUPP;
+	}
+
+	/*
+	 * Define which test result data should be written to the test result
+	 * JSON file.
+	 */
+	const struct json_entry aead_testresult_entries[] = {
+		/* also write empty CT (in case we only perform auth tests) */
+		{"ct",		{.data.buf = &vector->data, WRITER_BIN},	FLAG_OP_ENC | FLAG_OP_AFT},
+		/* also write empty PT (in case we only perform auth tests) */
+		{"pt",		{.data.buf = &vector->data, WRITER_BIN},	FLAG_OP_DEC | FLAG_OP_AFT},
+		{"testPassed",	{.data.integer = &vector->integrity_error, WRITER_BOOL_TRUE_TO_FALSE},	FLAG_OP_DEC | FLAG_OP_AFT},
+	};
+	const struct json_testresult aead_testresult = SET_ARRAY(aead_testresult_entries, callbacks);
+
+	/*
+	 * Define one particular test vector that is expected in the JSON
+	 * file. For example:
+	 * {
+         *   "tcId": 2171,
+         *   "key": "1529BAC6229586F057FAA59353851686",
+         *   "pt": "",
+         *   "aad": "4B11160620475D8EE440C3795CF62D26"
+         * },
+	 *
+	 * After parsing each individual test vector, the test should be
+	 * executed and the result should be written to a JSON file.
+	 */
+	const struct json_entry aead_test_entries[] = {
+		{"iv",		{.data.buf = &vector->iv, PARSER_BIN},		FLAG_OPTIONAL  | FLAG_OP_ENC | FLAG_OP_AFT},
+		{"key",		{.data.buf = &vector->key, PARSER_BIN}, 	FLAG_OP_ENC | FLAG_OP_AFT},
+		{"aad",		{.data.buf = &vector->assoc, PARSER_BIN}, 	FLAG_OP_ENC | FLAG_OP_AFT},
+		{"pt",		{.data.buf = &vector->data, PARSER_BIN}, 	FLAG_OP_ENC | FLAG_OP_AFT},
+
+		{"iv",		{.data.buf = &vector->iv, PARSER_BIN}, 		FLAG_OP_DEC | FLAG_OP_AFT},
+		{"key",		{.data.buf = &vector->key, PARSER_BIN},		FLAG_OP_DEC | FLAG_OP_AFT},
+		{"aad",		{.data.buf = &vector->assoc, PARSER_BIN},	FLAG_OP_DEC | FLAG_OP_AFT},
+		{"ct",		{.data.buf = &vector->data, PARSER_BIN},	FLAG_OP_DEC | FLAG_OP_AFT},
+	};
+	const struct json_array aead_test = SET_ARRAY(aead_test_entries, &aead_testresult);
+
+	const struct json_entry aead_testgroup_entries[] = {
+		{"payloadLen",	{.data.integer = &vector->ptlen, PARSER_UINT},	FLAG_OP_ENC | FLAG_OP_AFT},
+		{"tests",	{.data.array = &aead_test, PARSER_ARRAY},	FLAG_OP_ENC | FLAG_OP_AFT},
+
+		{"payloadLen",	{.data.integer = &vector->ptlen, PARSER_UINT},	FLAG_OP_DEC | FLAG_OP_AFT},
+		{"tests",	{.data.array = &aead_test, PARSER_ARRAY},	FLAG_OP_DEC | FLAG_OP_AFT}
+	};
+	const struct json_array aead_testgroup = SET_ARRAY(aead_testgroup_entries, NULL);
+
+	/*
+	 * Define the anchor of the tests in the highest level of the JSON
+	 * input data. For example:
+	 * {
+	 *   "acvVersion": "0.2",
+	 *   "vsId": 1564,
+	 *   "algorithm": "AES-GCM",
+	 *   "direction": "encrypt",
+	 *   "testGroups": [
+	 */
+	const struct json_entry aead_testanchor_entries[] = {
+		{"testGroups",	{.data.array = &aead_testgroup, PARSER_ARRAY}, 0 }
+	};
+	const struct json_array aead_testanchor = SET_ARRAY(aead_testanchor_entries, NULL);
+
+	/* Process all. */
+	return process_json(&aead_testanchor, "1.0", in, out);
 }
 
 static int gcm_tester(struct json_object *in, struct json_object *out,
@@ -197,6 +274,9 @@ static int gcm_tester(struct json_object *in, struct json_object *out,
 
 	memset(&vector, 0, sizeof(struct aead_data));
 	vector.cipher = cipher;
+
+	if (cipher == ACVP_GCMSIV)
+		return aead_siv_tester(in, out, &vector, &gcm_callbacks);
 
 	return aead_tester(in, out, &vector, &gcm_callbacks);
 }
@@ -324,7 +404,7 @@ static int ccm_tester(struct json_object *in, struct json_object *out,
 
 static struct cavs_tester gcm =
 {
-	ACVP_GCM,
+	ACVP_GCM | ACVP_GCMSIV,
 	gcm_tester,	/* process_req */
 	NULL
 };