+++ parser/parser_dh.c
@@ -237,7 +237,7 @@ static int dh_tester(struct json_object *in, struct json_object *out,
 	const struct json_array dh_testanchor = SET_ARRAY(dh_testanchor_entries, NULL);
 
 	/* Process all. */
-	return process_json(&dh_testanchor, "0.5", in, out);
+	return process_json(&dh_testanchor, "1.0", in, out);
 }
 
 static struct cavs_tester dh =