+++ parser/parser_dsa.c
@@ -412,7 +412,7 @@ static int dsa_tester(struct json_object *in, struct json_object *out,
 	const struct json_array dsa_testanchor = SET_ARRAY(dsa_testanchor_entries, NULL);
 
 	/* Process all. */
-	return process_json(&dsa_testanchor, "0.5", in, out);
+	return process_json(&dsa_testanchor, "1.0", in, out);
 }
 
 static struct cavs_tester dsa =