+++ parser/parser_ecdsa.c
@@ -325,7 +325,7 @@ static int ecdsa_tester(struct json_object *in, struct json_object *out,
 	const struct json_array ecdsa_testanchor = SET_ARRAY(ecdsa_testanchor_entries, NULL);
 
 	/* Process all. */
-	return process_json(&ecdsa_testanchor, "0.5", in, out);
+	return process_json(&ecdsa_testanchor, "1.0", in, out);
 }
 
 static struct cavs_tester ecdsa =