+++ parser/parser_eddsa.c
@@ -250,7 +250,7 @@ static int eddsa_tester(struct json_object *in, struct json_object *out,
 	const struct json_array eddsa_testanchor = SET_ARRAY(eddsa_testanchor_entries, NULL);
 
 	/* Process all. */
-	return process_json(&eddsa_testanchor, "0.5", in, out);
+	return process_json(&eddsa_testanchor, "1.0", in, out);
 }
 
 static struct cavs_tester eddsa =