+++ parser/parser_rsa.c
@@ -333,7 +333,7 @@ static int rsa_tester(struct json_object *in, struct json_object *out,
 	const struct json_array rsa_testanchor = SET_ARRAY(rsa_testanchor_entries, NULL);
 
 	/* Process all. */
-	return process_json(&rsa_testanchor, "0.5", in, out);
+	return process_json(&rsa_testanchor, "1.0", in, out);
 }
 
 static struct cavs_tester rsa =